[Windows] `flow ide` should exit gracefully if the server closes the socket
Summary:
On Linux (and OSX I assume), when the server dies or closes the
connection, the `flow ide` process will get an `End_of_file`. On Windows, it
gets an ECONNRESET and crashes. This fixes that.

I've also changed the exit code to something more standard. `No_server_running`
seemed like the closest exit code, since this almost always means that the
`flow server` has died.

Differential Revision: D5981193

fbshipit-source-id: 97525ce8738da83e3256033c80eb75e630f274d2