Fix the --timeout flag
Summary:
We've had a `--timeout` flag for awhile, but it's always been a
little bit broken, and only really worked for timing out connecting to
the server. Now that we connect instantly to the monitor, it really
doesn't work.

This fixes it by using the `Timeout` module to wrap
`connect_and_make_request`. Now the whole process of connecting to the
monitor, sending a request, and waiting for a response can time out
properly.

Differential Revision: D6628751

fbshipit-source-id: 32b741022114a1309dac38ed3c051edc784c93ce